Add SHARE_PROVIDER addon type and pull QR Code sharing out into a core addon that uses the new API
