+++ code/core/src/qrcode/constants.ts
@@ -1,0 +1,1 @@
+export const ADDON_ID = 'storybook/qrcode';
